pyproject.toml
[project]
name = "blanket-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "openai>=1.82.1",
    "prompt-toolkit>=3.0.51",
]

[project.scripts]
# package.module:func
blanketml= "blanketml:main"
[tool.pyright]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipython>=9.2.0",
    "numpy>=2.3.1",
]
